// packages/client-daemon-ts/src/models/BrowserOpenRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BrowserOpenRequest = {
    /**
     * Whether to use incognito mode
     */
    incognito?: boolean;
    /**
     * Whether to open the 9222 debug port
     */
    remoteDebug?: boolean;
    url?: string;
};